test: move some cases out of bucket_triggers
The test file was renamed from storage_2 in a previous commit to
get more trigger-related cases. Now the unrelated parts are moved
back to storage_2.

Part of #173

NO_DOC=refactoring
